refactor(gemini): Restructure examples into per-example directories
