+++ components/model/gemini/examples/generate_with_image/test.jpg
