+++ components/model/gemini/examples/image_generate/image_generate.go
@@ -1,0 +1,94 @@
+/*
+ * Copyright 2025 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package main
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	"google.golang.org/genai"
+
+	"github.com/cloudwego/eino-ext/components/model/gemini"
+	"github.com/cloudwego/eino/schema"
+)
+
+func main() {
+	apiKey := os.Getenv("GEMINI_API_KEY")
+
+	ctx := context.Background()
+	client, err := genai.NewClient(ctx, &genai.ClientConfig{
+		APIKey: apiKey,
+	})
+	if err != nil {
+		log.Fatalf("NewClient of gemini failed, err=%v", err)
+	}
+	defer func() {
+		if err != nil {
+			log.Printf("close client error: %v", err)
+		}
+	}()
+
+	cm, err := gemini.NewChatModel(ctx, &gemini.Config{
+		Client: client,
+		Model:  "gemini-2.5-flash-image-preview",
+		ResponseModalities: []gemini.GeminiResponseModality{
+			gemini.GeminiResponseModalityText,
+			gemini.GeminiResponseModalityImage,
+		},
+	})
+	if err != nil {
+		log.Fatalf("NewChatModel of gemini failed, err=%v", err)
+	}
+
+	/*
+		The generated multimodal content is stored in the `AssistantGenMultiContent` field.
+		For this example, the resulting message will have a structure similar to this:
+
+		resp := &schema.Message{
+			Role: schema.Assistant,
+			AssistantGenMultiContent: []schema.MessageOutputPart{
+				{
+					Type: schema.ChatMessagePartTypeImageURL,
+					Image: &schema.MessageOutputImage{
+						MessagePartCommon: schema.MessagePartCommon{
+							Base64Data: &base64String, // The base64 encoded image data
+							MIMEType:   "image/png",
+						},
+					},
+				},
+			},
+		}
+	*/
+	resp, err := cm.Generate(ctx, []*schema.Message{
+		{
+			Role: schema.User,
+			UserInputMultiContent: []schema.MessageInputPart{
+				{
+					Type: schema.ChatMessagePartTypeText,
+					Text: "Generate an image of a cat",
+				},
+			},
+		},
+	})
+	if err != nil {
+		log.Printf("Generate error: %v", err)
+		return
+	}
+	fmt.Printf("Assistant: %s\n", resp)
+}
