+++ components/model/gemini/examples/stream/stream.go
@@ -1,0 +1,90 @@
+/*
+ * Copyright 2025 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package main
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"log"
+	"os"
+
+	"google.golang.org/genai"
+
+	"github.com/cloudwego/eino-ext/components/model/gemini"
+	"github.com/cloudwego/eino/schema"
+)
+
+func main() {
+	apiKey := os.Getenv("GEMINI_API_KEY")
+
+	ctx := context.Background()
+	client, err := genai.NewClient(ctx, &genai.ClientConfig{
+		APIKey: apiKey,
+	})
+	if err != nil {
+		log.Fatalf("NewClient of gemini failed, err=%v", err)
+	}
+	defer func() {
+		if err != nil {
+			log.Printf("close client error: %v", err)
+		}
+	}()
+
+	cm, err := gemini.NewChatModel(ctx, &gemini.Config{
+		Client: client,
+		Model:  "gemini-2.5-flash",
+		ThinkingConfig: &genai.ThinkingConfig{
+			IncludeThoughts: true,
+			ThinkingBudget:  nil,
+		},
+	})
+	if err != nil {
+		log.Fatalf("NewChatModel of gemini failed, err=%v", err)
+	}
+	stream, err := cm.Stream(ctx, []*schema.Message{
+		{
+			Role:    schema.User,
+			Content: "Write a short poem about spring.",
+		},
+	})
+	if err != nil {
+		log.Printf("Stream error: %v", err)
+		return
+	}
+
+	fmt.Println("Assistant: ")
+	for {
+		resp, err := stream.Recv()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			log.Printf("Stream receive error: %v", err)
+			return
+		}
+
+		fmt.Println("frame: ")
+		if len(resp.Content) > 0 {
+			fmt.Println("content: ", resp.Content)
+		}
+		if len(resp.ReasoningContent) > 0 {
+			fmt.Printf("ReasoningContent: %s\n", resp.ReasoningContent)
+		}
+	}
+	fmt.Println()
+}
